quick fix

+++ src/main/java/com/knockout/reservations/ReservationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ReservationRepository extends CrudRepository<Reservation, Long> {
+public interface ReservationRepository extends CrudRepository<Reservation, Integer> {
 
     List<Reservation> findByCourtId(Integer courtId);
